requirements.txt
fastapi==0.123.0
paho_mqtt==2.1.0
sqladmin==0.22.0
SQLAlchemy==2.0.44
